feat: Using TDD create a Card Class and a Deck Class.
Add unittests to ensure that cards created with the Card Class have
str and repr methods.
Test that card instances have a rank. That their rank is one of the
thirteen possible options, and if not, that it raises a ValueError.
Test that card instances have a suit. That their suit is one of four
possible options, and if not, that it raises a ValueError.

Add a unittest to ensure that decks created with the Deck Class
store no cards initially. Test that the add_cards instance method
adds card objects to the deck 'cards' attribute.