feat: Create Class method on Card to return a standard 52 card list.
Add __eq__ method to Card to assert that two cards are equal
if they have the same rank and the same suit.
Test that the new standard 52 card list was created successfully
by asserting that that first card in the list is the 2 of Clubs
and the last card in the list is the Ace of Spades
and the length of the list is 52.