feat: Create Hand class.
Each Hand object receives and stores a list of cards.
The private _card_rank_counts property creates a dict of
the card ranks and the number of times each occurs.
The _ranks_with_count method returns a dict with the ranks whose
count equals the number provided.
best_rank method then uses 'if' blocks to return the correct
string depending on the length of the ranks_with_three_of_a_kind or
ranks_with_pairs dicts.