refactor: Restructure best_rank logic into separate methods.
Create separate instance methods which return True
if the given hand of cards matches the criteria.
Create a tuple of tuples property which stores the
rank validations from best to worst and their corresponding
instance methods.
In the existing best_rank method, use a 'for' loop to iterate
through the property, trying each method in turn
until 'True' is returned.
Then return the rank in string form.
This provides better separation of logic and is easier read/follow.