feat: Add logic to evaluate a Straight.
Add a rank_index attribute, based on the RANKS Class attribute, to each Card instance.
Use this new attribute within a custom __lt__ method to
define the less-than operator with respect to Card Objects.
Add tests to check that each Card knows its own rank,
that two cards can be sorted by rank and that a list of Card object,
when sorted, are ordered by rank ascending.

Create a copy of the cards list being passed to the Hand
and sort the list before saving as the self.cards attribute.
Change the test_receives_and_stores_cards unittest to reflect this newly sorted attribute.
Add 'Straight' tuple to the _rank_validations_from_best_to_worst property.
Add _straight instance method that checks whether the current hand has 5 cards.
and if so, converts the cards list into a list of their respective indices.
Compare this to a list generated using range() and the cards list starting and ending indices,
plus one, to check for a strictly-increasing list.
Add unittests to check that the best_rank instance method correctly recognises a 'Straight'
and that a hand with only two consecutive cards does not return a 'Straight'.