feat: Add logic to identify a Flush.
Add a new property to the Hand Class that generates a card_suit_counts dict.
Add a new 'Flush' tuple to the _rank_validations_from_best_to_worst property.
Use dict-comprehension syntax to generate a dict of suits that occur 5 or more times.
If the length of that dict is equal to 1, the Hand contains a flush, return True.
The unittest test_figures_out_flush_is_best_rank now passes.