feat: Add test and logic to identify a Full House.
Add unittest to return 'Full House' if the given hand has a pair and three-of-a-kind.
Add a 'Full House' tuple with name and _full_house instance method to the start of the
_rank_validations_from_best_to_worst property tuple.
Write a new _full_house method that calls the existing methods for _three_of_a_kind and _pair.
If both evaluate to true, the hand contains a Full House.
All unittests now pass.