feat: Add logic to recognise Hands with a Straight Flush or a Royal Flush.
Add instance method to the Hand Class to check for a Straight Flush.
Reuse existing _flush and _straight methods.

Add instance method to the Hand Class to check for a Royal Flush.
If the given hand returns True for a Straight Flush and if the final
card in the sorted cards has a rank of 'Ace' return True.