feat: Add Player Class.
Dependency Injection: Each Player Object is instantiated with a name
and a hand attribute. The hand attribute stores a Hand object.
The Player's best_hand instance method returns the result of a call to
the hand attribute/object's best_rank() method.

fix: Add _no_cards() instance method to Hand Class to return 'No Cards'
and not 'High Card' when .best_rank() is called on a hand containing an
empty list of cards.

fix: Return False within the _royal_flush instance method
if is_straight_flush evaluates to False as this means that
there are fewer than five cards in the hand. This prevents
an IndexError when the is_royal variable tries to index
into the empty list.