feat: Add Game Class.
Create Game Class with 'deck' and 'players' attributes.
A 'play' instance method calls 'shuffle' on the Deck object.
Within the Deck Class, the 'shuffle' instance method uses
the 'shuffle()' function from the 'ramdom' module to shuffle the
'_cards' attribute, which is a list.
Use the @patch decorator from unittest.mock to intercept the
call to random.shuffle to assert that it was called with the
list of cards.