feat: Add an instance method to remove cards from the Deck.
remove_cards() accepts an integer, which is the number of cards to be removed from the deck.
Using list slicing syntax, starting from the first element of the list
assign the cards to be removed to a variable, cards_to_remove.
Delete the specified number of cards from the _cards attribute.
Return the removed cards from the instance method using the cards_to_remove variable.

Within the associated unittest, use assertEqual to test the return value of
the remove_cards instance method and the altered state of the deck object
after it has been invoked.