refactor: Expanding Player and Hand Classes to Add New Cards.
Add the functionality to give the Player new cards.
Remove the instantiation of a Hand object with a 'cards' list.
Instantiate every Hand with an empty list.

Add an instance method 'add_cards' to the Player object.
This talks to their Hand and passes on the new cards.
The Hand takes the new cards, adds them to the existing collection of cards that it has,
sorts them in order of rank and then saves them to the 'cards' attribute.

Remove the cards lists from every instantiation of a Hand() within the testing suite.
Use the new add_cards() method to assign the list of cards to the Hand objects.
All tests now pass.