feat: Add Hand Technical Representation and Deck Length.
Add __repr__ method to the Hand class to return a
joined list of each Card's string representation.
This allows the user to call player.hand and see
a clear string representation of the cards in their hand.

Add __len__ method to the Deck class to return
the length of the Deck's 'cards' list when
len is called on any Deck.