feat: Add Logic to Deal the Flop Cards.
Within the Game Class, create a method '_deal_flop_cards'
that invokes the 'remove_cards' method on the Deck with
an argument of 3 and assigns these same three cards to
each player in the players list.

refactor: Add 'setUp' to GameTest Class.

Instantiate variables for the first_two_cards and
second_two_cards which point to lists of the Card instances.
Instantiate a community_cards variable and associated list of
three Cards for the three cards dealt during the Flop.
Remove the duplicated code from the existing GameTest Class methods.