refactor: Create PairValidator Class.
Move logic for validation of a 'Pair' to new PairValidator Class.
Create 'is_valid' and 'valid_cards' methods, so its structure
mimic the other validator classes.
Instantiate a PairValidator object within the
'_rank_validations_from_best_to_worst' tuple in hand.py.
Remove the '_pair' method on the Hand Class.

fix: Instantiate another PairValidator object within the '_full_house'
method. This method's logic had relied on the original '_pair' method.
Instead, invoke the 'is_valid' method on the PairValidator object to return a Boolean
and remove the error.

fix: Add an 'if' block to the __lt__ method within card.py.
If cards are of equivalent ranks, then evaluate their suit in alphabetical order.
This prevents errors when pairs of cards of equivalent rank were returned in
the order in which they were added to the cards list, which could not be predetermined.