refactor: Create ThreeOfAKindValidator Class.
Move logic for validation of Three of a Kind to the new class
Make the class consistent with the other validators by adding
the 'is_valid' and 'valid_cards' instance methods.
Import and instantiate the new class within the Hand class.
Remove the '_three_of_a_kind' method from the class.

fix: Replace the call to the '_three_of_a_kind' method
within the _full_house method with a new instantiated
ThreeOfAKindValidator object and invoke the 'is_valid' method
on it to return the correct Boolean value.