refactor: Create StraightFlushValidator Class.
Move the logic relating to Straight Flushes to a the new class.
This includes the same '_collections_of_five_straight_cards_in_a_row' property
and '_every_element_increasing_by_one' method that the Straight Class also
uses. Create a new FiveCardRanksInARowValidator Class to house this logic
and make the Straight Class and StraightFlush Class inherit it from the superclass.
Add a new property to the StraightFlushValidator Class '_straight_flush_card_batches'.
This property returns the five card batch for every batch which has a confirmed straight
only if, after removing duplicates using set comprehension syntax, only one
suit remains in the set. Now, even if a Straight and a Flush exist within the
collection of cards, only those cards that make up a StraightFlush will be returned.

If more than one StraightFlush is present, within the 'valid_cards' method,
the list-slicing syntax ensures that the higher of the two is returned.